Give the three fake arm states one definition
`test_harness.py`, `test_golden_pipeline.py` and `test_ds_writer_agent.py` each
carried a `roboarm.State` fake over the same four properties. They now share
`FakeRobotState` in `drivers/roboarm/tests/fakes.py`, beside the interface it
fakes, with `make_robot_state` for the stationary identity-rotation case the two
policy tests want.

The merged fake copies on read, which the golden producer relied on and the
other two are indifferent to. The golden file is unchanged.
